refactor: style: fix: refactor: separate out flask routes and dash callbacks refactor: consolidate dash layouts refactor: removed unused dash components fix: renamed duplicated callback functions style: linted, pylint 10/10
