+++ ncov19_dash/components/stats_table.py
@@ -1,9 +1,6 @@
 import gc
 import requests
 import pandas as pd
-import dash_bootstrap_components as dbc
-import dash_html_components as html
-import dash_table
 from ncov19_dash.utils import REVERSE_STATES_MAP
 from ncov19_dash import config
 
@@ -15,15 +12,20 @@ def stats_table(state="US"):
     URL = config.NCOV19_API + config.COUNTY
     try:
         response = requests.get(URL)
-    except Exception as ex:
+    except ValueError as ex:
         print(f"[ERROR] stats_table error accessing ncov19.us API, {ex}")
-        data = {"state_name": "john", "county_name": "cena", "confirmed": 0, "death": 0}
+    else:
+        data = {"state_name": "john",
+                "county_name": "cena",
+                "confirmed": 0, "death": 0}
 
     if response.status_code == 200:
         data = response.json()["message"]
         data = pd.DataFrame.from_records(data)
     else:
-        data = {"state_name": "john", "county_name": "cena", "confirmed": 0, "death": 0}
+        data = {"state_name": "john",
+                "county_name": "cena",
+                "confirmed": 0, "death": 0}
 
     if state in ["US", "United States"]:
         data = data.groupby(["state_name"])[["confirmed", "death"]].sum()
