+++ ncov19_dash/layout/desktop_footer.py
