+++ ncov19_dash/layout/desktop_nav.py
@@ -1,5 +1,4 @@
 import dash_bootstrap_components as dbc
-import dash_core_components as dcc
 import dash_html_components as html
