+++ ncov19_dash/layout/mobile_about.py
