+++ ncov19_dash/layout/mobile_footer.py
