+++ ncov19_dash/layout/mobile_nav.py
