--- ncov19_dash/pages/__init__.py
@@ -1,6 +1,0 @@
-from .nav import navbar
-from .mobile_nav import mobile_navbar
-from .footer import footer
-from .mobile_footer import mobile_footer
-from .about import about_body
-from .mobile_about import mobile_about_body
