+++ ncov19_dash/routes/__init__.py
@@ -1,0 +1,2 @@
+from .sitemap import sitemap
+from .robots import robots
