+++ ncov19_dash/utils/__init__.py
@@ -1,7 +1,1 @@
-import os
-# from .config import *
-from .settings import *
-
-
-
-# print(f"[DEBUG] Config being used is: {config.__class__.__name__}")
+from .data import *
