--- ncov19_dash/wsgi.py
@@ -1,4 +1,0 @@
-from ncov19_dash import server
-
-if __name__ == "__main__":
-    server.run(debug=True, port=8050)
